requirements.txt
Django==4.0.3    
django-cors-headers==3.11.0
djangorestframework==3.13.1 
djangorestframework-simplejwt==5.1.0
Pillow==9.0.1
mysqlclient==2.1.0
django-rest-auth